Revamp player traits to support instance based adjustments
Prior to this change, the traits were just dependent on the player class in a static manner. Now, the traits can be adjusted per player which is important when the player reaches the stage that buffs his traits.
It is now possible to define which traits are being buffed at that time in the Proficiency class.